+++ packages/@expo/config-plugins/src/ios/XcodeProjectFile.ts
@@ -73,25 +73,3 @@ export function createBuildSourceFile({
   }
   return project;
 }
-
-/**
- * Read a source file from the Xcode project
- *
- * @param nativeProjectRoot absolute path to the native app root `user/app/ios`
- * @param filePath path relative to the `nativeProjectRoot` for the file to read `user/app/ios/myapp/foobar.swift`
- */
-export function readBuildSourceFile({
-  project,
-  nativeProjectRoot,
-  filePath,
-}: {
-  project: XcodeProject;
-  nativeProjectRoot: string;
-  filePath: string;
-}): string | null {
-  const absoluteFilePath = path.join(nativeProjectRoot, filePath);
-  if (!fs.existsSync(absoluteFilePath)) {
-    return null;
-  }
-  return fs.readFileSync(absoluteFilePath, { encoding: 'utf8' });
-}