+++ packages/expo-updates/cli/src/resolveRuntimeVersion.ts
@@ -3,6 +3,7 @@ import chalk from 'chalk';
 
 import { Command } from './cli';
 import { requireArg, assertArgs, getProjectRoot } from './utils/args';
+import { CommandError } from './utils/errors';
 import * as Log from './utils/log';
 
 export const resolveRuntimeVersion: Command = async (argv) => {
@@ -42,14 +43,19 @@ Resolve expo-updates runtime version
 
   const platform = requireArg(args, '--platform');
   if (!['ios', 'android'].includes(platform)) {
-    throw new Error(`Invalid platform argument: ${platform}`);
+    throw new CommandError(`Invalid platform argument: ${platform}`);
   }
 
   const debug = args['--debug'];
 
-  const runtimeVersionInfo = await resolveRuntimeVersionAsync(getProjectRoot(args), platform, {
-    silent: true,
-    debug,
-  });
+  let runtimeVersionInfo;
+  try {
+    runtimeVersionInfo = await resolveRuntimeVersionAsync(getProjectRoot(args), platform, {
+      silent: true,
+      debug,
+    });
+  } catch (e: any) {
+    throw new CommandError(e.message);
+  }
   console.log(JSON.stringify(runtimeVersionInfo));
 };